feat(3.5): 添加字段

+++ lease/lease-service/src/main/java/cn/cuiot/dmp/lease/dto/charge/ChargeManagerInsertVo.java
@@ -37,7 +37,7 @@ public class ChargeManagerInsertVo {
      * 收费标准 0自定义金额
      */
     @NotNull(message = "收费标准不能为空")
-    private Byte chargeStandard;
+    private Long chargeStandard;
 
     /**
      * 应收金额/本金
