+++ lease/lease-service/src/main/java/cn/cuiot/dmp/lease/entity/charge/TbChargePlain.java
@@ -38,7 +38,7 @@ public class TbChargePlain extends YjBaseEntity {
      * 收费标准 0自定义金额
      */
     @TableField(value = "charge_standard")
-    private Byte chargeStandard;
+    private Long chargeStandard;
 
     /**
      * 应收金额/本金
